fix some tests
