+++ src/data/data_spillter.py
@@ -1,6 +1,4 @@
 import pandas as pd
-from data_loader import load_data
-from data_cleaner import clean_data
 
 
 # TODO: Move this to somewhere else (Don't think it is being used anywhere now but can be useful in the future)
@@ -49,16 +47,62 @@ def split_patient_data_by_day(patients_dfs: pd.DataFrame, patient_id: str) -> di
     return daily_dfs
 
 
-if __name__ == "__main__":
-    patient_id = "p01"
-    df = clean_data(load_data(data_source_name="kaggle_brisT1D", dataset_type="train"))
+# TODO: Remove the dependency of p_num. Kaggle data is the very few dataset where there are multiple patients in the same file.
+def get_train_validation_split(
+    df: pd.DataFrame,
+    num_validation_days: int = 20,
+) -> tuple[pd.DataFrame, pd.DataFrame]:
+    """
+    Split the data into train and validation sets based on complete days,
+    where a day is defined as 6am-6am.
 
-    # Test the function with sample data
-    daily_dfs = split_patient_data_by_day(df, "p01")
+    Args:
+        df (pd.DataFrame): Input dataframe with datetime column
+        num_validation_days (int): Number of complete 6am-6am days to use for validation
 
-    # Print results
-    print(f"Number of days: {len(daily_dfs)}")
-    n = 3
-    key = f"p01_{n}"
-    print(f"\nDay {n} data:")
-    print(daily_dfs[key])
+    Returns:
+        tuple[pd.DataFrame, pd.DataFrame]: (train_data, validation_data) where validation
+            contains exactly num_validation_days of complete 6am-6am days
+    """
+    if "datetime" not in df.columns:
+        raise ValueError(
+            "datetime column not found in data. Please run create_datetime_index first."
+        )
+
+    df["datetime"] = pd.to_datetime(df["datetime"])
+
+    # For each patient:
+    # 1. Find the last 6am timestamp
+    # 2. Go back num_validation_days to get the start of validation
+    # 3. Trim any data after the last 6am
+    validation_data_list = []
+    train_data_list = []
+
+    for _, patient_df in df.groupby("p_num"):
+        # Get timestamps where hour is 6 (6am)
+        six_am_times = patient_df[patient_df["datetime"].dt.hour == 6]["datetime"]
+
+        if len(six_am_times) == 0:
+            continue
+
+        # Get the last 6am timestamp
+        last_six_am = six_am_times.max()
+
+        # Calculate the start of validation period (num_validation_days before last_six_am)
+        validation_start = last_six_am - pd.Timedelta(days=num_validation_days)
+
+        # Split the patient's data
+        patient_validation = patient_df[
+            (patient_df["datetime"] >= validation_start)
+            & (patient_df["datetime"] <= last_six_am)
+        ]
+        patient_train = patient_df[patient_df["datetime"] < validation_start]
+
+        validation_data_list.append(patient_validation)
+        train_data_list.append(patient_train)
+
+    # Combine all patients' data
+    validation_data = pd.concat(validation_data_list)
+    train_data = pd.concat(train_data_list)
+
+    return train_data, validation_data
